Move LatexUI to Components and remove unused UI.
