+++ apps/web/src/lib/components/Latex.svelte
