--- packages/eslint-config-custom/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "eslint-config-custom",
-  "version": "0.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-turbo": "latest",
-    "eslint-plugin-svelte3": "^4.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
