--- packages/ui/index.ts
@@ -1,6 +1,0 @@
-export { default as Container } from './components/Container.svelte';
-export { default as Formula } from './components/Formula.svelte';
-export { default as LatexUI } from './components/Latex.svelte';
-export { default as RoundButton } from './components/RoundButton.svelte';
-export { default as ToggleFullscreen } from './components/ToggleFullscreen.svelte';
-export { default as UI } from './components/UI.svelte';
