--- packages/ui/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "ui",
-  "version": "0.0.0",
-  "description": "Styles and components for use in svelte websites",
-  "type": "module",
-  "module": "index.ts",
-  "svelte": "index.ts",
-  "license": "MIT",
-  "scripts": {
-    "lint": "prettier --check --ignore-path=../../.prettierignore . && eslint \".\"",
-    "format": "prettier --write --ignore-path=../../.prettierignore ."
-  },
-  "devDependencies": {
-    "@types/katex": "^0.14.0",
-    "eslint-config-custom": "*",
-    "svelte": "^3.44.0",
-    "utils": "*"
-  },
-  "dependencies": {
-    "katex": "^0.16.4",
-    "screenfull": "^6.0.2"
-  }
-}
